[리팩토링] 읽기 전용 서비스 레이어 분리 #5
