+++ seminar1/src/main/java/org/example/service/PostRetrieveService.java
@@ -1,0 +1,34 @@
+package org.example.service;
+
+import jakarta.persistence.EntityNotFoundException;
+import java.util.List;
+import lombok.RequiredArgsConstructor;
+import org.example.domain.Category;
+import org.example.domain.Post;
+import org.example.dto.response.PostGetResponse;
+import org.example.repository.PostJpaRepository;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+@Service
+@Transactional(readOnly = true)
+@RequiredArgsConstructor
+public class PostRetrieveService {
+    private final CategoryService categoryService;
+    private final PostJpaRepository postJpaRepository;
+
+    public List<PostGetResponse> getPosts(Long memberId) {
+        return postJpaRepository.findAllByMemberId(memberId)
+                .stream()
+                .map(post -> PostGetResponse.of(post, getCategoryByPost(post)))
+                .toList();
+    }
+
+    public PostGetResponse getById(Long postId) {
+        Post post = postJpaRepository.findById(postId).orElseThrow(() -> new EntityNotFoundException("해당하는 게시글이 없습니다."));
+        return PostGetResponse.of(post, getCategoryByPost(post));
+    }
+    private Category getCategoryByPost(Post post) {
+        return categoryService.getByCategoryId(post.getCategoryId());
+    }
+}
